pyproject.toml
[project]
name = "bratbot"
version = "0.1.0"
description = "A bratty, condescending Discord bot powered by a self-hosted LLM via Ollama"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord.py>=2.6,<3.0",
    "httpx>=0.28,<1.0",
    "redis>=5.2,<6.0",
    "pydantic-settings>=2.7,<3.0",
    "structlog>=24.4,<26.0",
    "PyNaCl>=1.5,<2.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9,<1.0",
    "watchfiles>=1.0,<2.0",
    "pytest>=8.3,<9.0",
    "pytest-asyncio>=0.24,<1.0",
]
model = [
    "fastapi>=0.110,<1.0",
    "uvicorn>=0.29,<1.0",
    "pydantic>=2.6,<3.0",
    "httpx>=0.28,<1.0",
    "PyNaCl>=1.5,<2.0",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bratbot", "src/bonniebot", "src/common"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "model"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
